feat: Move `module` package into `service` package
