+++ cli/commands/catalog/service/module/doc_test.go
