+++ cli/commands/catalog/service/module/testdata/find_modules/gitdir/config
