+++ cli/commands/catalog/service/module/testdata/find_modules/modules/eks-aws-auth-merger/README.adoc
